+++ Godeps/_workspace/src/github.com/onsi/ginkgo/CHANGELOG.md
@@ -1,3 +1,13 @@
+## HEAD
+
+Improvements:
+
+- `Skip(message)` can be used to skip the current test.
+
+Bug Fixes:
+
+- Ginkgo tests now fail when you `panic(nil)` (#167)
+
 ## 1.2.0 5/31/2015
 
 Improvements